backend/requirements.txt
django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
pymongo
python-dotenv==1.2.1
gunicorn==23.0.0
